+++ packages/eyes-webdriverio-5/src/WDIOCheckSettings.js
@@ -1,0 +1,5 @@
+const {DriverCheckSettings} = require('@applitools/eyes-sdk-core')
+const WrappedElement = require('./WDIOWrappedElement')
+const Frame = require('./WDIOFrame')
+
+module.exports = DriverCheckSettings.specialize({WrappedElement, Frame})