+++ packages/eyes-webdriverio-5/src/WDIOFrame.js
@@ -1,0 +1,4 @@
+const {Frame} = require('@applitools/eyes-sdk-core')
+const WrappedElement = require('./WDIOWrappedElement')
+
+module.exports = Frame.specialize({WrappedElement})