+++ packages/eyes-webdriverio-5/test/coverage/index.js
@@ -140,9 +140,9 @@ function initialize() {
             }, ${isFully})`,
           )
         : result.storeCommand(
-            `await eyes.checkRegion(By.css('${target}'), ${matchTimeout}, ${
+            `await eyes.checkRegionBy(By.css('${target}'), ${
               tag ? '"' + tag + '"' : undefined
-            })`,
+            }, ${matchTimeout}, ${isFully})`,
           )
     } else {
       result.storeCommand(`{`)
@@ -233,11 +233,15 @@ function initialize() {
   }
 
   async function switchToFrame(selector) {
-    result.storeCommand(`await driver.switchToFrame(By.css('${selector}'))`)
+    result.storeCommand(`
+      await driver.switchToFrame(
+        await driver.findElement('css selector', '${selector}')
+      )
+    `)
   }
 
   async function type(selector, text) {
-    result.storeCommand(`await driver.elementSendKeys('${selector}', '${text}')`)
+    result.storeCommand(`await driver.$('${selector}').then(el => el.setValue('${text}'))`)
   }
   return {
     hooks: {