+++ packages/eyes-webdriverio-5/test/it/TestEdgeSwitchToFrame.spec.js
@@ -14,7 +14,7 @@ describe('TestEdgeSwitchToFrame', () => {
           local: 'false',
           seleniumVersion: '3.5.2',
         },
-        resolution: '1920x1080',
+        // resolution: '1920x1080',
         browserName: 'Edge',
         browserVersion: '18.0',
       },